[msvc] added python cleanup options

+++ bin/cleanup.sh
@@ -42,10 +42,13 @@ while (($#)); do
             install="--clean"
             ;;
 
+        "--python") # cleans up python cache folders
+            python="--python"
+            ;;
+
         "--build") # build project using cmake
             build="--build"
             ;;
-
 
         "--help") # shows command description
             help
@@ -71,7 +74,13 @@ if [[ "${install}" == "--clean" ]]; then
     rm -rf "${pwd}/lib"
     rm -rf "${pwd}/logs"
     rm -f "${pwd}/src/std/version.h"
+    find "${pwd}" -type f -name "callgrind.out.*" -delete
+    find "${pwd}" -type f -name "*.s" -delete
+fi
+
+if [[ "${python}" == "--python" ]]; then
     find "${pwd}" -type d -name "__pycache__" -exec rm -rf {} +
+    find "${pwd}" -type d -name ".pytest_cache" -exec rm -rf {} +
 fi
 
 if [[ "${build}" == "--build" ]]; then
